+++ NEWS.md
@@ -1,6 +1,7 @@
 # nexus 0.1.0.9000
 ## New classes and methods
 * Add `split()` methods to divide into groups.
+* Add `hist()` methods to plot ILR histograms.
 
 # nexus 0.1.0